activation: Move the code a bit
This will allow adding subcommands for the activation handler.
